Add MANIFEST.in.
This allows generating an sdist as follows:
```
git clean -xdf  # Repo must be clean, we make no effort to not include build artefacts!
python setup.py sdist
```
One can check that the sdist is valid by converting it to a wheel with
`pip wheel --no-deps dist/libasd-VERSION.tar.gz` and checking that the
wheel installs correctly.